Added changes to homewor_2 for the third task
