+++ homework_2/additionalTask/Main.java
@@ -1,5 +1,7 @@
 import fields.TicTac;
 import results.Results;
+import xml.SaveToXML;
+
 import java.io.IOException;
 
 public class Main {
@@ -7,5 +9,20 @@ public class Main {
         TicTac ticTac = new TicTac();
         ticTac.game();
         new Results();
+        new AddDataGameToXml().appendChild();
+        SaveToXML.getTransformer();
+        SaveToXML.XmlToConsole();
+    }
+}
+
+class AddDataGameToXml{
+    void appendChild(){
+        SaveToXML.root.appendChild(SaveToXML.player1);
+        SaveToXML.root.appendChild(SaveToXML.player2);
+        SaveToXML.root.appendChild(SaveToXML.game);
+        for(int i = 0; i < 9; i++) SaveToXML.game.appendChild(SaveToXML.listStep.get(i));
+        SaveToXML.root.appendChild(SaveToXML.gameResult);
+        SaveToXML.gameResult.appendChild(SaveToXML.winPlayer);
+        SaveToXML.doc.appendChild(SaveToXML.root);
     }
 }
