+++ homework_2/additionalTask/results.txt
@@ -1,2 +1,3 @@
-Результат человека: 5
+Результат человека: 1
 Результат бота: 0
+Вы сыграли: 1
