--- homework_3/additionalTask/entities/Entity.java
@@ -1,6 +1,0 @@
-package entities;
-
-public interface Entity {
-    void turn();
-    boolean isCellValid(int x, int y);
-}
