--- homework_3/additionalTask/info.xml
@@ -1,19 +1,0 @@
-<?xml version="1.0" encoding="windows-1251" standalone="no"?>
-<GamePlay>
-    <Player id="1" name="Alexsandr" symbol="X"/>
-    <Player id="2" name="Al_Bot" symbol="O"/>
-    <Game>
-        <step num="1" playerId="1">1 1</step>
-        <step num="2" playerId="2">1 3</step>
-        <step num="3" playerId="1">1 2</step>
-        <step num="4" playerId="2">2 3</step>
-        <step num="5" playerId="1">3 3</step>
-        <step num="6" playerId="2">2 2</step>
-        <step num="7" playerId="1">3 1</step>
-        <step num="8" playerId="2">3 2</step>
-        <step num="9" playerId="1">2 1</step>
-    </Game>
-    <GameResult>
-        <Player id="1" name="Alexsandr" symbol="X"/>
-    </GameResult>
-</GamePlay>
