--- homework_3/additionalTask/results.txt
@@ -1,3 +1,0 @@
-Результат человека: 1
-Результат бота: 0
-Вы сыграли: 1
